[REF] plugins: split geo chart data in two
This commit splits the geo_features plugin into two separate plugins:
geo_loader and geo_features.
The geo_loader plugin is responsible for loading the geo chart data,
which is needed for the evaluation.
The geo_features plugin is responsible for providing the geo chart stuff
for actions in the UI, such as the geo chart region selector.

Task: 6428159
